+++ src/clj/rems/db/applications.clj
@@ -286,6 +286,7 @@
   (let [application (get-application app-id)]
     (assert (= :application.state/draft (:application/state application))
             (str "Tried to delete application " app-id " which is not a draft!")))
+  (db/delete-application-attachments! {:application app-id})
   (db/delete-application-events! {:application app-id})
   (db/delete-application! {:application app-id})
   (reload-cache!))