+++ CHANGELOG.md
@@ -12,9 +12,12 @@ Changes since v2.17
 
 ### Fixes
 - Errors for invalid inputs (field values that are too long, invalid email addresses, etc.) are now rendered nicely. Previously the applicant just saw a "Save draft: Failed" message. (#2611)
+- The application page no longer jumps to the top after adding an attachment. (#2616)
+- Deleting drafts with attachments now works.
 - A handler is now considered a handler even before first application comes in.
 
 ### Additions
+- REMS now supports PostgreSQL version 13. (#2642)
 
 ## v2.17 "Isokaari" 2021-04-12